The most significant changes include the update of the target framework for two projects, the removal of code in two files, the addition of new files with new code, and the addition of several unit tests for the `SenderProcessTests` class.
1. The target framework for the projects `DelegateActionAndFuncAppSample.csproj` and `TestDelegatesActionFunc.csproj` has been updated from `net7.0` to `net8.0`. This change ensures that the projects are compatible with the latest .NET framework.

2. The entire code in `Processing.cs` and `ProcessingTests.cs` has been removed. This change might be due to the code in these files being obsolete or no longer needed.

3. New files `EmailSender.cs`, `MessageMap.cs`, `Messages.cs`, `SenderProcess.cs`, `SendNotifications.cs`, `SmsSender.cs`, `TypeMessageEnum.cs`, and `WhatsAppSender.cs` have been added with new code. This new code includes classes for sending different types of messages (Email, SMS, WhatsApp), a class for mapping message data, a class for processing messages, and an enumeration for message types.

4. Several unit tests for the `SenderProcessTests` class have been added. These tests are designed to verify the functionality of the `SenderProcess` class methods, including `ProcessListMessages`, `ProcessFile`, `Processed`, and the `Send` methods of `WhatsAppSender`, `SmsSender`, and `EmailSender` classes. The tests use the Moq framework to create mock objects of `SendNotifications` and `SenderProcess` classes, which are used to isolate the unit of code under test. The tests cover various scenarios and use assertions to verify that the actual outcomes match the expected outcomes.
