+++ csharp-advanced-topics/DelegatesActionAndFuncInCsharp/DelegatesActionAndFunc/Messages.cs
@@ -1,0 +1,10 @@
+﻿namespace DelegateFuncAppSample;
+
+public record Messages
+{
+    public required int TypeMessage;
+    public required string Recipient;
+    public required string Message;
+    public string? Email;
+    public string? PhoneNumber;
+}
