+++ csharp-advanced-topics/DelegatesActionAndFuncInCsharp/DelegatesActionAndFunc/SenderProcess.cs
@@ -1,92 +1,6 @@
 ﻿using CsvHelper;
-using CsvHelper.Configuration;
-using Xunit;
 
 namespace DelegateFuncAppSample;
-
-public enum TypeMessageEnum
-{
-    WhatsApp = 0,
-    Sms = 1,
-    Email = 2
-}
-
-public record Messages
-{
-    public required int TypeMessage;
-    public required string Recipient;
-    public required string Message;
-    public string? Email;
-    public string? PhoneNumber;
-}
-
-public sealed class MessageMap : ClassMap<Messages>
-{
-    public MessageMap()
-    {
-        Map(m => m.TypeMessage).Index(0);
-        Map(m => m.Recipient).Index(1);
-        Map(m => m.Message).Index(2);
-        Map(m => m.Email).Index(3);
-        Map(m => m.PhoneNumber).Index(4);
-    }
-}
-
-public class SendNotifications
-{
-    public Func<Messages, Action> WhatsApp { get; } = messages => () =>
-    {
-        if (messages.TypeMessage != (int)TypeMessageEnum.WhatsApp) throw new Exception("Type mismatch");
-
-        Console.WriteLine($"Sending WhatsApp message to {messages.Recipient}: {messages.Message}");
-    };
-
-    public Func<Messages, Action> Sms { get; } = messages => () =>
-    {
-        if (messages.TypeMessage != (int)TypeMessageEnum.Sms) throw new Exception("Type mismatch");
-
-        Console.WriteLine($"Sending SMS to {messages.Recipient}: {messages.Message}");
-    };
-
-    public Func<Messages, Action> Email { get; } = messages => () =>
-    {
-        if (messages.TypeMessage != (int)TypeMessageEnum.Email) throw new Exception("Type mismatch");
-
-        Console.WriteLine($"Sending Email to {messages.Recipient}: {messages.Message}");
-    };
-
-    public int[] CountTypes { get; set; } = new[] { 0, 0, 0 };
-}
-
-
-public class WhatsAppSender
-{
-    public void Send(Messages messages)
-    {
-        if (messages.TypeMessage != (int)TypeMessageEnum.WhatsApp) throw new Exception("Type mismatch");
-
-        Console.WriteLine($"Sending WhatsApp message to {messages.Recipient}: {messages.Message}");
-    }
-}
-
-public class SmsSender
-{
-    public void Send(Messages messages)
-    {
-        if (messages.TypeMessage != (int)TypeMessageEnum.Sms) throw new Exception("Type mismatch");
-
-        Console.WriteLine($"Sending SMS to {messages.Recipient}: {messages.Message}");
-    }
-}
-public class EmailSender
-{
-    public void Send(Messages messages)
-    {
-        if (messages.TypeMessage != (int)TypeMessageEnum.Email) throw new Exception("Type mismatch");
-
-        Console.WriteLine($"Sending Email to {messages.Recipient}: {messages.Message}");
-    }
-}
 
 public class SenderProcess
 {
