+++ csharp-advanced-topics/DelegatesActionAndFuncInCsharp/DelegatesActionAndFunc/TypeMessageEnum.cs
@@ -1,0 +1,8 @@
+﻿namespace DelegateFuncAppSample;
+
+public enum TypeMessageEnum
+{
+    WhatsApp = 0,
+    Sms = 1,
+    Email = 2
+}
